requirements.txt
Flask~=3.1.1
openai~=1.84.0
python-dotenv~=1.1.0
flask-cors~=6.0.0
httpx~=0.28.1